requirements.txt
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brotli==1.0.9
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
cloud-init==22.3.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
FlightRadarAPI==1.2.1
gsconfig-py3==1.0.8
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pexpect==4.8.0
Pillow==9.4.0
ptyprocess==0.7.0
pushbullet.py==0.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.21
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-debian===0.1.43ubuntu1
python-magic==0.4.27
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
rgbmatrix @ file:///mnt/c/Users/markd/PycharmProjects/LEDispi/rpi-rgb-led-matrix/bindings/python
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.3
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
ubuntu-advantage-tools==27.11.2
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
websocket-client==1.4.2
zipp==1.0.0
zope.interface==5.4.0
